+++ markdown/gui/src/events/ListenerDemo.java
@@ -1,25 +1,48 @@
 package events;
 
-import javax.swing.JFrame;
-import javax.swing.SwingUtilities;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import java.util.logging.Logger;
+import javax.swing.JButton;
+import javax.swing.JPanel;
 
-/** Rahmen für die Listener-Demos */
-public class ListenerDemo {
-    /** Erzeuge die Komponenten in neuem EDT-Job statt im main()-Tread */
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(
-                new Runnable() {
-                    public void run() {
-                        JFrame frame = new JFrame("Listener Demo");
-                        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+/** Demo: Umgang mit mehreren Listenern/Observables */
+public class ListenerDemo extends JPanel {
+    private final JButton singleButton;
+    private final JButton multiButton;
+    private static final Logger LOG = Logger.getLogger(ListenerDemo.class.getName());
 
-                        // frame.setContentPane(new MouseListenerDemo());
-                        // frame.setContentPane(new MouseAdapterDemo());
-                        frame.setContentPane(new ListenerPanel());
+    /** Erzeuge ein Listener-Panel */
+    public ListenerDemo() {
+        super();
 
-                        frame.pack();
-                        frame.setVisible(true);
-                    }
-                });
+        singleButton = new JButton("single listener");
+        add(singleButton);
+        multiButton = new JButton("multi listener");
+        add(multiButton);
+
+        registerSameListenerWithMultipleObservables();
+        registerMultipleListenersWithSameObservable();
+    }
+
+    private void registerSameListenerWithMultipleObservables() {
+        // ein Listener kann bei mehreren Observables registriert werden
+        Handler single = new Handler("Wuppie");
+        singleButton.addActionListener(single);
+        multiButton.addActionListener(single);
+    }
+
+    private void registerMultipleListenersWithSameObservable() {
+        // ein Observable kann mehrere Listener bedienen
+        multiButton.addActionListener(new Handler("Foo"));
+        multiButton.addActionListener(new Handler("Bar"));
+    }
+
+    /** @param s Referenz auf Text-Widget, welches in actionPerformed() verändert wird */
+    private record Handler(String s) implements ActionListener {
+        @Override
+        public void actionPerformed(ActionEvent e) {
+            LOG.info("Button: '" + e.getActionCommand() + "', Handler: '" + s + "'");
+        }
     }
 }
